refactor: split acp translator presentation helpers
